Change condition whether to check other browsers
Instead of ignoring other browsers if the cookie jar is empty, continue
checking other browsers if the cookie jar does not have the required
number of elements (specifically 2; one for "LEETCODE_SESSION" and one
for "csrftoken"). This is necessary because when
`browser_cookie3.chrome()` is called the "LEETCODE_SESSION" cookie will
be missing and the program would previously give up instead of examining
the other browsers.